requirements.txt
redis==6.0.0
pyyaml==6.0.2